pyproject.toml
[tool.poetry]
name = "bottelegramgpt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-telegram-bot = "^20.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"